Moved oauth to /v1/auth endpoint, implemented expiring tokens
