+++ api/resources_portal/views/__init__.py
@@ -1,5 +1,6 @@
 from resources_portal.views.address import AddressViewSet
 from resources_portal.views.attachment import AttachmentViewSet, local_file_view
+from resources_portal.views.auth import AuthViewSet
 from resources_portal.views.document_views import (
     MaterialDocumentView,
     OrganizationDocumentView,
